Scroll on new lines
This addition will make sure that the document scrolls properly when inserting a new line. You might want to make it optional and have a setting for it. I am using this by modifying the installed extension code directly, and it works like a charm, but it's not tested properly as I haven't tried compiling the ts files.